chore - rename skeleton file
